scripts/requirements.txt
azure-data-tables>=12.4.0
azure-identity>=1.15.0
